lmadani/requirements.txt
Flask==2.3.0
pycryptodome==3.17.0